requirements.txt
rasa==3.6.2
requests==2.31.0
pyyaml>=5.3.1,<6.0
python-dotenv==1.0.0